[Test] Fix warnings from PVS-Studio

+++ src/internal/cbufferoutstream.cpp
@@ -61,7 +61,7 @@ STDMETHODIMP CBufferOutStream::Write( const void* data, UInt32 size, UInt32* pro
     }
 
     auto oldPos = ( mCurrentPosition - mBuffer.begin() );
-    const size_t newPos = static_cast< size_t >( oldPos ) + size;
+    const size_t newPos = static_cast< size_t >( oldPos ) + static_cast< size_t >( size );
     if ( newPos > mBuffer.size() ) {
         try {
             mBuffer.resize( newPos );
